Optionally check for direct link support.
This is incredibly annoying. There's no info on the user object that indicates
whether or not direct links are supported. The closest you could maybe get is
inferring based on the total quota size, but that would still be wrong for some
enterprise users. And the share request won't fail if the user can't use
direct links, because technically the share still succeeded, it just won't
work with the direct download URL that they so happily provide.

So, during the information gathering, we create a new file, share that file,
and then try to download that file. If it works, then direct links are
supported. Regardless, we delete the test file.

We need this info, hence the hoop jumping, but it adds a lot of overhead that
others may not need. So this extra check is disabled by default and can be
enabled with an option during the provider configuration.